Improve structure and documentation of Stats connector
