+++ Sources/InstantSearchCore/Stats/Connector/StatsConnector.swift
@@ -38,44 +38,16 @@ public class StatsConnector {
 
 }
 
-public extension StatsConnector {
-  
-  /**
-    - Parameters:
-      - searcher: Searcher that handles your searches
-      - interactor: Logic applied to the stats
-      - controller: Controller that interfaces with a concrete stats view
-   */
-  convenience init<Controller: StatsTextController>(searcher: SingleIndexSearcher,
-                                               interactor: StatsInteractor = .init(),
-                                               controller: Controller) {
-    self.init(searcher: searcher,
-              interactor: interactor)
-    connectController(controller)
-  }
-  
-  /**
-   Establishes a connection with the controller
-   - Parameters:
-     - controller: Controller that interfaces with a concrete stats view
-   - Returns: Established connection
-  */
-  @discardableResult func connectController<Controller: StatsTextController>(_ controller: Controller) -> some Connection {
-    let connection = interactor.connectController(controller)
-    controllerConnections.append(connection)
-    return connection
-  }
-  
-}
-
 extension StatsConnector: Connection {
   
   public func connect() {
     searcherConnection.connect()
+    controllerConnections.forEach { $0.connect() }
   }
 
   public func disconnect() {
     searcherConnection.disconnect()
+    controllerConnections.forEach { $0.disconnect() }
   }
   
 }
